requirements.txt
streamlit>=1.30.0
google-generativeai>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
